[Update] first phase prototype implemented
Now, the project have a tflite trained model to identify some coins on a photo. The model however is too simple and was trained with a small dataset of images, thus been inacurrate for the most cases, but has enought to test the "test.jpeg" image (was completlly new for the model) really well!